// frontend/src/app/shared/models/certificate-overview/PaginatedCertificateOverviewResponse.ts
import { CertificateOverviewResponse } from "./CertificateOverviewResponse";

export interface PaginatedCertificateOverviewResponse {
  content: CertificateOverviewResponse[];
  pageable: {
    pageNumber: number;
    pageSize: number;
    sort: {
      empty: boolean;
      sorted: boolean;
      unsorted: boolean;
    };
    offset: number;
    paged: boolean;
    unpaged: boolean;
  };
  last: boolean;
  totalElements: number;
  totalPages: number;
  size: number;
  number: number;
  first: boolean;
  numberOfElements: number;
  sort: {
    empty: boolean;
    sorted: boolean;
    unsorted: boolean;
  };
  empty: boolean;
}