// frontend/src/app/shared/models/certificate-overview/CertificateOverviewResponse.ts
export interface CertificateOverviewResponse {
  id: number;
  serialNumber: string;
  commonName: string;
  organizationName: string;
  organizationalUnit: string | null;
  country: string | null;
  email: string | null;
  status: 'ACTIVE' | 'REVOKED' | 'EXPIRED';
  validFrom: string;
  validTo: string;
  keyAlgorithm: string | null;
  keySize: number | null;
  signatureAlgorithm: string | null;
  revocationReason: string | null;
  revokedAt: string | null;
  SAN?: string | null;
  includeSubjectKeyIdentifier?: boolean;
  includeAuthorityKeyIdentifier?: boolean;
  includeExtendedKeyUsage?: boolean;
  hasPrivateKey : boolean;
}